1. 部分组件完善 2. i18n完善

+++ fx-plus/src/main/java/cn/oyzh/fx/plus/controls/digital/ByteTextField.java
@@ -16,7 +16,7 @@ public class ByteTextField extends NumberTextField {
         super(unsigned);
     }
 
-    public ByteTextField(boolean unsigned, Integer maxLen) {
+    public ByteTextField(boolean unsigned, Long maxLen) {
         super(unsigned, maxLen);
     }
