+++ fx-plus/src/main/java/cn/oyzh/fx/plus/controls/digital/ShortTextField.java
@@ -16,7 +16,7 @@ public class ShortTextField extends NumberTextField {
         super(unsigned);
     }
 
-    public ShortTextField(boolean unsigned, Integer maxLen) {
+    public ShortTextField(boolean unsigned, Long maxLen) {
         super(unsigned, maxLen);
     }
